Split parser.* into hash.*, value.* and parser.*
