+++ src/celengine/dsodb.cpp
@@ -26,6 +26,7 @@
 #include "parseobject.h"
 #include "multitexture.h"
 #include "meshmanager.h"
+#include "tokenizer.h"
 #include <celutil/debug.h>
 
 #include <celengine/galaxy.h>
