+++ src/celengine/meshmanager.cpp
@@ -21,6 +21,7 @@
 #include "texmanager.h"
 #include "meshmanager.h"
 #include "modelgeometry.h"
+#include "tokenizer.h"
 
 #include <cel3ds/3dsread.h>
 #include <celmodel/modelfile.h>
