+++ src/celengine/virtualtex.cpp
@@ -20,6 +20,7 @@
 #include <celcompat/filesystem.h>
 #include <celutil/filetype.h>
 #include "parser.h"
+#include "tokenizer.h"
 #include "virtualtex.h"
