+++ src/celestia/destination.cpp
@@ -13,6 +13,7 @@
 #include <celutil/util.h>
 #include <celengine/astro.h>
 #include <celengine/parser.h>
+#include <celengine/tokenizer.h>
 #include "destination.h"
 
 using namespace std;
